+++ src/components/projects/dashboard/project-manager/fotos/carousel/FotosInterno.tsx
@@ -9,6 +9,7 @@ import 'swiper/css/pagination';
 import 'swiper/css/scrollbar';
 import { deletePrismaProjectImage } from "@/app/services/deleteInvestmentImage";
 import { FotosInternoInput } from "@/components/CreateProjects/Inputs/FotosInternoInput";
+import axios from "axios";
 
 
 interface FotosDestaquesProps {
@@ -34,7 +35,19 @@ export function FotosInterno({ projectData, openImage }: FotosDestaquesProps) {
     useEffect(() => {
 
         const deleteImage = async (imageID: Investment["images"][0]["id"]) => {
+
             try {
+
+                const imageToDelete = interno.find((image) => image.id === imageID);
+
+                if (!imageToDelete) {
+                    console.warn(`Imagem com ID ${imageID} não encontrada no array destaques`);
+                    return;
+                }
+
+                const responseImgDeleted = await axios.post('/api/delete-image', { 
+                    imageUrl: imageToDelete.url 
+                  });
 
                 const response = await deletePrismaProjectImage(projectData.id, imageID)
                 // Salvando alterações no estado
@@ -120,7 +133,7 @@ export function FotosInterno({ projectData, openImage }: FotosDestaquesProps) {
                                 >
                                     <Image
                                         cursor={editMode ? 'pointer' : 'grabbing'}
-                                        src={`/assets/projects/${img.url}`}
+                                        src={`${img.url}`}
                                         h={190}
                                         w='100%'
                                         objectFit={'cover'}
