+++ src/components/projects/dashboard/project-manager/plantas/carousel/FotosPlanta.tsx
@@ -9,6 +9,7 @@ import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import 'swiper/css/scrollbar';
+import axios from "axios";
 
 
 interface FotosPlantaProps {
@@ -36,7 +37,20 @@ export function FotosPlanta({ projectData, openImage }: FotosPlantaProps) {
     useEffect(() => {
 
         const deleteImage = async (imageID: Investment["images"][0]["id"]) => {
+
             try {
+
+                const imageToDelete = planta.find((image) => image.id === imageID);
+
+                if (!imageToDelete) {
+                    console.warn(`Imagem com ID ${imageID} não encontrada no array destaques`);
+                    return;
+                }
+
+                const responseImgDeleted = await axios.post('/api/delete-image', { 
+                    imageUrl: imageToDelete.url 
+                  });
+
 
                 const response = await deletePrismaProjectImage(projectData.id, imageID)
                 // Salvando alterações no estado
@@ -122,7 +136,7 @@ export function FotosPlanta({ projectData, openImage }: FotosPlantaProps) {
                                 >
                                     <Image
                                         cursor={editMode ? 'pointer' : 'grabbing'}
-                                        src={`/assets/projects/${img.url}`}
+                                        src={`${img.url}`}
                                         h={190}
                                         w='100%'
                                         objectFit={'cover'}
